+++ packages/in_app_purchase/in_app_purchase_storekit/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.3.6+1
+
+* Updates links for the merge of flutter/plugins into flutter/packages.
+
 ## 0.3.6
 
 * Updates minimum Flutter version to 3.3 and iOS 11.