+++ packages/video_player/video_player_android/CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 2.3.11
 
+* Updates links for the merge of flutter/plugins into flutter/packages.
 * Updates minimum Flutter version to 3.0.
 * Bumps Exoplayer dependency versions to 2.18.3